moved auth service into subfolder
